inblock tweaks
git-svn-id: svn+ssh://astrometry.net/svn/trunk/projects/tractor@22666 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ wise3.py
@@ -1081,7 +1081,7 @@ def stage104(opt=None, ps=None, tractor=None, band=None, bandnum=None, T=None,
 	R.set(band, np.array([src.getBrightness().getBand(band) for src in cat]))
 	R.set(band + '_var', vflux)
 	R.row = S.row
-	R.inblock = S.inblock
+	R.inblock = S.inblock.astype(np.uint8)
 
 	imstats = tabledata()
